+++ riverdriver/riverdatabasesql/river_database_sql_driver.go
@@ -241,7 +241,7 @@ func (e *Executor) JobInsertFastMany(ctx context.Context, params []*riverdriver.
 		insertJobsParams.ScheduledAt[i] = scheduledAt
 		insertJobsParams.State[i] = string(params.State)
 		insertJobsParams.Tags[i] = strings.Join(tags, ",")
-		insertJobsParams.UniqueKey[i] = sliceutil.DefaultIfEmpty(params.UniqueKey, nil)
+		insertJobsParams.UniqueKey[i] = sliceutil.FirstNonEmpty(params.UniqueKey)
 		insertJobsParams.UniqueStates[i] = pgtypealias.Bits{Bits: pgtype.Bits{Bytes: []byte{params.UniqueStates}, Len: 8, Valid: params.UniqueStates != 0}}
 	}
 
@@ -315,6 +315,7 @@ func (e *Executor) JobInsertFull(ctx context.Context, params *riverdriver.JobIns
 	job, err := dbsqlc.New().JobInsertFull(ctx, e.dbtx, &dbsqlc.JobInsertFullParams{
 		Attempt:      int16(min(params.Attempt, math.MaxInt16)), //nolint:gosec
 		AttemptedAt:  params.AttemptedAt,
+		AttemptedBy:  params.AttemptedBy,
 		Args:         string(params.EncodedArgs),
 		CreatedAt:    params.CreatedAt,
 		Errors:       sliceutil.Map(params.Errors, func(e []byte) string { return string(e) }),